requirements.txt
gradio>=4.19.0
pypdf>=4.0.0
pymupdf>=1.24.3
pillow